Move dummy Add preprocessor to the proper location
