+++ orangecontrib/snom/widgets/preprocessors/__init__.py
@@ -1,3 +1,4 @@
 # load and register editors
 from . import registration_example  # noqa: F401
 from . import phase_unwrap  # noqa: F401
+from . import multiply  # noqa: F401
