+++ orangecontrib/snom/widgets/preprocessors/registry.py
@@ -1,0 +1,5 @@
+from orangecontrib.spectroscopy.widgets.preprocessors.registry import (
+    PreprocessorEditorRegistry,
+)
+
+preprocess_image_editors = PreprocessorEditorRegistry()
